requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
openai==1.58.1
pyyaml==6.0.2
httpx==0.28.1
pytest==8.3.4
# Tier 1 CUAD loader + F2/Jaccard scoring
datasets>=2.14
numpy>=1.24
